Updated API resources

+++ database/migrations/2020_02_15_141803_create_individuals_table.php
@@ -14,12 +14,13 @@ class CreateIndividualsTable extends Migration
     public function up()
     {
         Schema::create('individuals', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('sacco_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('sacco_id');
             $table->string('first_name');
             $table->string('last_name');
             $table->string('email');
             $table->string('gender');
+            //$table->foreign('sacco_id')->references('id')->on('saccos');
         });
     }
