+++ database/migrations/2020_02_15_144120_create_transactions_table.php
@@ -14,10 +14,11 @@ class CreateTransactionsTable extends Migration
     public function up()
     {
         Schema::create('transactions', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('individual_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('individual_id');
             $table->integer('amount');
             $table->enum('type', ['deposit', 'withdrawal', 'transfer']);
+            //$table->foreign('individual_id')->references('id')->on('individuals');
         });
     }
